refactor OEM Icon download to take place in lockscreen manager
